requirements.txt
Pillow==10.0.1
altgraph==0.17.4
pyinstaller==4.10
pyinstaller-hooks-contrib==2023.11
